Make dotnet-nightly workaround version-agnostic
If the `Debug/dnxcore50` path does not exist, don't worry about it.

+++ build.ps1
@@ -12,7 +12,7 @@ cd ../..
 cd src/Microsoft.PowerShell.Linux.Host
 dotnet publish --framework dnxcore50 --output $BIN
 # Temporary fix for dotnet publish
-cp $BIN/Debug/dnxcore50/* $BIN
+if (Test-Path $BIN/Debug/dnxcore50) { cp $BIN/Debug/dnxcore50/* $BIN }
 # Copy files that dotnet-publish does not currently deploy
 cp *_profile.ps1 $BIN
 cd ../..
